examples: add library version to scoped db names. (#445)
This updates the web and intro examples to scope the local database name
to the library version. The result is that users who previously had a
local database with a previous library version will use a new database
after the library is updated.

The intention is to prevent some local-db incompatibility bugs with
users that return to the website.

Note that the implementation uses a json file import. Is there any
potential compatibility issue with this?