fix(electric): Fix column name quoting in one of the trigger functions (#474)
Fixes VAX-1062.

I had to refactor the way the affected SQL function is defined in order
for it to work both when running Electric for the first time and when
upgrading from an earlier version. I recommend reviewing this PR one
commit at a time to follow the progression of function code repackaging.