fix (client): solve migration bug with indexes (#471)
This PR solves the index bug described in #457.
The bug was caused by the client migration code which tried to build
table triggers even if the table information is missing. The code was
modified not to build any triggers if the table information is missing.
The table information may be missing e.g. when only creating an index.