fix (generator): correctly capitalise table named model (#515)
This PR fixes the bug for a table named "model", described by a user on
[Discord](https://discord.com/channels/933657521581858818/933657523049885698/1156342498453639260).
We were parsing the Prisma schema in order to capitalise model names but
if the model name itself was "model" we would capitalise the "model"
keyword instead of the model name "model". This is fixed by replacing
the exact match in the string instead of replacing the first occurrence
of the table name.

Also extended the unit test to check this case.