feat(starter): starter improvements to avoid port clashes (#505)
This PR improves the starter to enable several (independent) Electric
applications to run on the same machine and thus fixes #497 and #489.
Also addressed VAX-1114.

The changes are as follows:
- Avoid volume clashes by putting project name on the app name
- Avoid web server port clashes by supporting a custom port to be used
and fix it everywhere in the template
- Avoid Electric port clashes by supporting a custom port to be used and
fix it everywhere in the template
- Introduced 2 modes for the starter: fast mode and interactive mode
- Fast mode: app name and ports (optional) are provided as command line
arguments
- Interactive mode: no arguments to the script, app name and ports are
asked via prompts
- Changed start to check that the chosen port for Electric is free. If
not, prompt user if they want to change the port
- Script to change the project’s configuration with other Electric and
web server ports. Exposed as a new `ports:configure` command.
- Changed `backend:start` to check for port clashes and if there is a
port clash suggest to run `ports:configure`.
- Changed `yarn start` and `client:generate` to check that the Electric
port the app is configured to use is the app's own Electric service
(e.g. catches problem where project 1 is running Electric on port 5133
and another project is accidentally configured to also connect to port
5133).
- Modified build script to use forward requests to right esbuild server.
This fixes a bug where if an esbuild server was already running, the new
app would forward its requests to the esbuild server of the other app.

tldr: it should now be possible to run several Electric applications.
Port clashes should always be detected and reported with a suggestion to
reconfigure the app with other ports.

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>